Update with interface and utils files
Refactor file structure
