+++ src/lib/utils/getInitials.ts
@@ -1,0 +1,7 @@
+export const getInitials = (name = '') =>
+  name
+    .split(' ')
+    .map((part) => part[0])
+    .splice(0, 2)
+    .join('')
+    .toUpperCase();
